put handler logic in the controllers
