+++ src/routes/index.ts
@@ -2,6 +2,7 @@ import { Router } from 'express';
 import v1 from './v1';
 
 const router = Router();
-router.use(v1);
+
+router.use('/v1', v1);
 
 export default router;
