+++ src/routes/v1/index.ts
@@ -2,6 +2,7 @@ import { Router } from 'express';
 import chats from './chats';
 
 const router = Router();
-router.use(chats);
+
+router.use('/chats', chats);
 
 export default router;
